+++ src/status_im/ui/screens/db.cljs
@@ -114,7 +114,7 @@
 
 (spec/def :navigation.screen-params/usage-data vector?)
 
-(spec/def :navigation.screen-params/display-collectible map?)
+(spec/def :navigation.screen-params/collectibles-list map?)
 
 (spec/def :navigation/screen-params (spec/nilable (allowed-keys :opt-un [:navigation.screen-params/network-details
                                                                          :navigation.screen-params/browser
@@ -124,7 +124,7 @@
                                                                          :navigation.screen-params/edit-contact-group
                                                                          :navigation.screen-params/dapp-description
                                                                          :navigation.screen-params/usage-data
-                                                                         :navigation.screen-params/display-collectible])))
+                                                                         :navigation.screen-params/collectibles-list])))
 
 (spec/def :desktop/desktop (spec/nilable any?))