+++ src/status_im/ui/screens/wallet/collectibles/subs.cljs
@@ -1,6 +1,11 @@
 (ns status-im.ui.screens.wallet.collectibles.subs
   (:require [re-frame.core :as re-frame]))
 
-(re-frame/reg-sub :collectibles
-                  (fn [db [_ s]]
-                    (vals (get-in db [:collectibles s]))))
+(re-frame/reg-sub :collectibles :collectibles)
+
+(re-frame/reg-sub
+ :screen-collectibles
+ :<- [:collectibles]
+ :<- [:get-screen-params]
+ (fn [[collectibles {:keys [symbol]}]]
+   (mapv #(assoc (second %) :id (first %)) (get collectibles symbol))))