+++ src/status_im/ui/screens/wallet/events.cljs
@@ -14,8 +14,6 @@
             status-im.ui.screens.wallet.request.events
             [status-im.constants :as constants]
             [status-im.ui.screens.navigation :as navigation]
-            [status-im.ui.screens.wallet.collectibles.views :as collectibles]
-            [status-im.utils.ethereum.erc721 :as erc721]
             [status-im.utils.money :as money]))
 
 (defn get-balance [{:keys [web3 account-id on-success on-error]}]
@@ -282,22 +280,3 @@
    {:db (-> db
             (assoc-in [:wallet :send-transaction] {})
             (navigation/navigate-back))}))
-
-(handlers/register-handler-fx
- :wallet/show-collectibles
- (fn [{:keys [db]} [_ i address {:keys [symbol] :as m}]]
-   (assoc (collectibles/load-collectibles-fx (:web3 db) symbol i address)
-          :dispatch [:navigate-to :display-collectible m])))
-
-(re-frame/reg-fx
- :load-collectibles
- (fn [[web3 symbol i address]]
-   (dotimes [n i]
-     (erc721/token-of-owner-by-index web3 (:address (tokens/symbol->token :mainnet symbol)) address n
-                                     #(re-frame/dispatch [:load-collectible symbol (.toNumber %2)])))))
-
-(handlers/register-handler-fx
- :load-collectible
- (fn [{db :db} [_ symbol id]]
-   (assoc (collectibles/load-collectible-fx symbol id)
-          :db db)))