+++ src/status_im/ui/screens/wallet/views.cljs
@@ -11,9 +11,9 @@
             [status-im.ui.screens.wallet.styles :as styles]
             [status-im.ui.screens.wallet.utils :as wallet.utils]
             [status-im.utils.money :as money]
-            status-im.ui.screens.wallet.collectibles.cryptokitties
-            status-im.ui.screens.wallet.collectibles.cryptostrikers
-            status-im.ui.screens.wallet.collectibles.etheremon))
+            status-im.ui.screens.wallet.collectibles.etheremon.views
+            status-im.ui.screens.wallet.collectibles.cryptostrikers.views
+            status-im.ui.screens.wallet.collectibles.cryptokitties.views))
 
 (defn toolbar-view []
   [toolbar/toolbar {:style styles/toolbar :flat? true}
@@ -88,11 +88,12 @@
   [list/item-icon {:icon      :icons/forward
                    :icon-opts {:color :gray}}])
 
-(defn- render-collectible [address-hex {:keys [symbol name icon amount] :as m}]
-  (let [i        (money/to-fixed amount)
-        details? (pos? i)]
-    [react/touchable-highlight (when details?
-                                 {:on-press #(re-frame/dispatch [:wallet/show-collectibles i address-hex m])})
+(defn- render-collectible [address-hex {:keys [symbol name icon amount] :as collectible}]
+  (let [items-number (money/to-fixed amount)
+        details?     (pos? items-number)]
+    [react/touchable-highlight
+     (when details?
+       {:on-press #(re-frame/dispatch [:show-collectibles-list address-hex collectible])})
      [react/view {:style styles/asset-item-container}
       [list/item
        [list/item-image icon]
@@ -100,8 +101,9 @@
         [react/text {:style               styles/asset-item-value
                      :number-of-lines     1
                      :ellipsize-mode      :tail
-                     :accessibility-label (str (-> symbol clojure.core/name clojure.string/lower-case) "-collectible-value-text")}
-         (or i 0)]
+                     :accessibility-label (str (-> symbol clojure.core/name clojure.string/lower-case)
+                                               "-collectible-value-text")}
+         (or items-number "...")]
         [react/text {:style           styles/asset-item-currency
                      :number-of-lines 1}
          name]]