+++ src/status_im/utils/ethereum/tokens.cljs
@@ -1,6 +1,7 @@
 (ns status-im.utils.ethereum.tokens
   (:require-macros [status-im.utils.ethereum.macros :refer [resolve-icons]])
-  (:require [clojure.string :as string]))
+  (:require [clojure.string :as string]
+            [status-im.utils.config :as config]))
 
 (defn- asset-border [color]
   {:border-color color :border-width 1 :border-radius 32})
@@ -386,21 +387,18 @@
                     :address  "0x9B11EFcAAA1890f6eE52C6bB7CF8153aC5d74139"
                     :decimals 8
                     :hidden?  true}
-                   {:symbol   :CK
-                    :nft?     true
-                    :name     "CryptoKitties"
-                    :address  "0x06012c8cf97bead5deae237070f9587f8e7a266d"
-                    :hidden?  true}
-                   {:symbol   :EMONA
-                    :nft?     true
-                    :name     "Etheremon"
-                    :address  "0xB2c0782ae4A299f7358758B2D15dA9bF29E1DD99"
-                    :hidden?  true}
-                   {:symbol   :STRK
-                    :nft?     true
-                    :name     "CryptoStrikers"
-                    :address  "0xdcaad9fd9a74144d226dbf94ce6162ca9f09ed7e"
-                    :hidden?  true}])
+                   {:symbol  :CK
+                    :nft?    true
+                    :name    "CryptoKitties"
+                    :address "0x06012c8cf97bead5deae237070f9587f8e7a266d"}
+                   {:symbol  :EMONA
+                    :nft?    true
+                    :name    "Etheremon"
+                    :address "0xB2c0782ae4A299f7358758B2D15dA9bF29E1DD99"}
+                   {:symbol  :STRK
+                    :nft?    true
+                    :name    "CryptoStrikers"
+                    :address "0xdcaad9fd9a74144d226dbf94ce6162ca9f09ed7e"}])
    :testnet
    (resolve-icons :testnet
                   [{:name     "Status Test Token"