+++ test/cljs/status_im/test/runner.cljs
@@ -43,6 +43,7 @@
             [status-im.test.utils.prices]
             [status-im.test.utils.keychain.core]
             [status-im.test.utils.universal-links.core]
+            [status-im.test.utils.http]
             [status-im.test.ui.screens.events]
             [status-im.test.ui.screens.accounts.login.events]))
 
@@ -98,5 +99,6 @@
  'status-im.test.utils.prices
  'status-im.test.utils.keychain.core
  'status-im.test.utils.universal-links.core
+ 'status-im.test.utils.http
  'status-im.test.ui.screens.events
  'status-im.test.ui.screens.accounts.login.events)